CI checks workflow (#205)
* Add mock keychain fixture.

* Copy mock keychain.

* Add task to generate test key.

* Update workflow.

* Fix binary in Cargo.toml.

* Fix tests on linux.

* Update workflow name.

* Use bash shell for windows.

* Enable RUST_BACKTRACE in checks.

* Fix libp2p deprecated behavior.

* Add caching to workflows.

* List keychains directory.

* Disable linux and windows for now.

* Configure mock keychain.

* Restore checks.

* Add --release to task to workaround windows stack overflow.

See https://github.com/clap-rs/clap/issues/5134.

* Beef up the macos runner.

* Debug test failures.

* Debug test failure in CI.

* Add more debug statements.

* Workaround for rename on temporary files.

On Linux renaming from /tmp is a different file system so it fails with
ErrorKind::CrossDevices but it's a nightly only variant so we can't add
a rename_or_copy() function yet.

For now just copy when using temp files.

* Do not specify write() to OpenOptions.

* Trying to figure out failing tests in CI.

* Print read dir entries in test.

* Hack for failing unit tests in CI.

* Fix hack for missing target directory.

* Bump patch version.

* Do not compile rexpect on windows.

* Tidy debug statements.

* Bump patch version.

* Enable linux in checks workflow, use apple silicon.

* Update test time variable for CI.

Allow for it to take a bit longer in CI.

* Use fail-fast, revert to x86_64 for apple.

+++ workspace/sdk/src/events/log/mod.rs
@@ -115,18 +115,30 @@ mod test {
 
     async fn mock_event_log_server_client(
     ) -> Result<(EventLogFile, EventLogFile, SecretId)> {
+        
+        // Required for CI which is setting the current 
+        // working directory to the workspace member rather
+        // than using the top-level working directory
+        if !vfs::try_exists("target").await? {
+            vfs::create_dir("target").await?;
+        }
+
         let server_file =
             PathBuf::from("target/mock-event-log-server.event_log");
         let client_file =
             PathBuf::from("target/mock-event-log-client.event_log");
-        let _ = vfs::remove_file(&server_file).await;
-        let _ = vfs::remove_file(&client_file).await;
+        if vfs::try_exists(&server_file).await? {
+            let _ = vfs::remove_file(&server_file).await;
+        }
+        if vfs::try_exists(&client_file).await? {
+            let _ = vfs::remove_file(&client_file).await;
+        }
 
         let vault: Vault = Default::default();
         let vault_buffer = encode(&vault).await?;
 
         let (id, data) = mock_secret().await?;
-
+        
         // Create a simple event log
         let mut server = EventLogFile::new(&server_file).await?;
         server
@@ -196,7 +208,10 @@ mod test {
     async fn event_log_diff() -> Result<()> {
         let partial =
             PathBuf::from("target/mock-event-log-partial.event_log");
-        let _ = vfs::remove_file(&partial).await;
+
+        if vfs::try_exists(&partial).await? {
+            let _ = vfs::remove_file(&partial).await;
+        }
 
         let (mut server, client, id) = mock_event_log_server_client().await?;
